+++ src/lib.rs
@@ -1,5 +1,6 @@
+pub mod add;
 mod change_type;
-mod changelog;
+pub mod changelog;
 pub mod cli;
 pub mod cli_config;
 pub mod config;